requirements.txt
langchain-ollama
langchain-openai
langchain-core
langchain-community
pynput
pywin32
rapidfuzz
Pillow
numpy
python-dotenv
google-generativeai
ollama
mem0ai
pyautogui
uiautomation
chromadb
neo4j
open-interpreter
langchain-neo4j
rank-bm25